Extract the load custom code function to separated file.
