+++ src/utils/load.js
@@ -1,0 +1,13 @@
+import vm from 'vm';
+
+export default function load(code) {
+  const exports = {};
+  const sandbox = {
+    exports,
+    module: { exports },
+  };
+
+  vm.runInNewContext(code, sandbox);
+
+  return sandbox.module.exports;
+}
